better card layouts, fix wallet disconnect redirect
